\name{precision-methods}
\docType{methods}
\alias{precision-methods}
\alias{precision,classifierOutput,character-method}
\alias{precision,classifierOutput,missing-method}
\alias{precision,classifierOutput,numeric-method}
\alias{precision}
\alias{recall-methods}
\alias{recall,classifierOutput,character-method}
\alias{recall,classifierOutput,missing-method}
\alias{recall,classifierOutput,numeric-method}
\alias{recall}
\alias{macroF1-methods}
\alias{macroF1,classifierOutput,character-method}
\alias{macroF1,classifierOutput,missing-method}
\alias{macroF1,classifierOutput,numeric-method}
\alias{macroF1}

\title{ Assessing classifier preformance }

\description{
  Methods for function \code{precision}, \code{recall} and
  \code{macroF1} in package \pkg{MLInterfaces}. 
}

\section{Methods}{
  \describe{
    \item{\code{signature(obj = "classifierOutput", type = "character")}}{
      %%  ~~describe this method here~~
    }
    \item{\code{signature(obj = "classifierOutput", type = "missing")}}{
      %%  ~~describe this method here~~
    }
    \item{\code{signature(obj = "classifierOutput", type = "numeric")}}{
      %%  ~~describe this method here~~
    }
  }
}


\keyword{methods}
