\name{xvalLoop-methods}
\docType{methods}
\alias{xvalLoop-methods}
\alias{xvalLoop,ANY-method}
\title{Cross-validation in clustered computing environments}
\description{Use cross-validation in a clustered computing environment}
\section{Methods}{
  See \code{\link{xvalLoop}} and the vignette for more information.
  \describe{
    \item{cluster = "ANY"}{This is the default method, returning the
      function \code{\link{lapply}} to be used in applying the
      cross-validation.}
}}
\keyword{methods}
