\name{MLIntInternals}
 \alias{getGrid}
 \alias{getGrid,data.frame-method}
 \alias{getGrid,ExpressionSet-method}
\alias{MLLabel }
\alias{MLOutput }
\alias{MLScore }
\alias{SOMBout }
\alias{classifOutput }
\alias{clustOutput }
\alias{groupIndex}
\alias{membMat }
\alias{probArray }
\alias{probMat }
\alias{qualScore }
\alias{silhouetteVec }
\alias{somout}

\title{MLInterfaces infrastructure}
\description{
These functions are internal tools for \code{MLInterfaces}.
Users will generally not call these functions directly.
}
\usage{
getGrid(x)
}

\arguments{
  \item{x}{a vector or matrix or ExpressionSet}
}
\details{
Forthcoming.
}
\value{
Functions with `new' as prefix are constructor helpers.
}

\author{VJ Carey <stvjc@channing.harvard.edu>}

\keyword{ classif }

